feat: 基本完成角色管理基础增删查改逻辑

+++ jiiiiiin-module-mngauth/src/main/java/cn/jiiiiiin/module/mngauth/controller/RoleController.java
@@ -1,6 +1,7 @@
 package cn.jiiiiiin.module.mngauth.controller;
 
 
+import cn.jiiiiiin.module.common.entity.mngauth.Admin;
 import cn.jiiiiiin.module.common.entity.mngauth.Resource;
 import cn.jiiiiiin.module.common.entity.mngauth.Role;
 import cn.jiiiiiin.module.common.enums.common.ChannelEnum;
@@ -9,6 +10,10 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.api.R;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.apache.commons.lang3.math.NumberUtils;
@@ -21,6 +26,7 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
+import java.util.List;
 
 /**
  * <p>
@@ -48,15 +54,16 @@ public class RoleController extends BaseController {
         return R.ok(roleService.page(new Page<Role>(current, size), new QueryWrapper<Role>().eq(Role.CHANNEL, channel).eq(Role.AUTHORITY_NAME, authorityName)));
     }
 
-    /**
-     * @param role 其中可能包含{@link Role#getResources()}，因为前端控件的原因，下面将会添加资源对应的父节点到需要关联的记录中
-     * @return
-     */
-    @PostMapping
-    public R<Role> create(@RequestBody Role role) {
+    @GetMapping("{id}")
+    public R<Role> searchRoleAndRelationRecords(@PathVariable Long id) {
+        return R.ok(roleService.getRoleAndRelationRecords(id));
+    }
+
+    private Long[] parseResourceIds(@NonNull Role role){
         val resourceIdsSet = new HashSet<Long>();
         role.getResources().forEach(resource -> {
-            for (String pid : resource.getPids().split(",")) {
+            val pids = resource.getPids().split(",");
+            for (String pid : pids) {
                 if (NumberUtils.isCreatable(pid)) {
                     val temp = Long.valueOf(pid);
                     if (!temp.equals(Resource.IS_ROOT_MENU)) {
@@ -67,13 +74,35 @@ public class RoleController extends BaseController {
             resourceIdsSet.add(resource.getId());
         });
         Long[] resourceIds = new Long[resourceIdsSet.size()];
-        roleService.save(role, resourceIdsSet.toArray(resourceIds));
+        return resourceIdsSet.toArray(resourceIds);
+    }
+
+    /**
+     * @param role 其中可能包含{@link Role#getResources()}，因为前端控件的原因，下面将会添加资源对应的父节点到需要关联的记录中
+     * @return
+     */
+    @PostMapping
+    public R<Role> create(@RequestBody Role role) {
+        roleService.save(role, parseResourceIds(role));
+        return success(role);
+    }
+
+    @PutMapping
+    public R<Role> update(@RequestBody Role role) {
+        roleService.update(role, parseResourceIds(role));
         return success(role);
     }
 
     @DeleteMapping
-    public R<Boolean> del(ArrayList<Long> idList) {
-        return success(roleService.remove(idList));
+    public R<Boolean> del(@RequestBody Admin data) {
+        val roles = data.getRoles();
+        if (roles.size() > 0) {
+            val idList = new HashSet<Long>();
+            roles.forEach(item -> idList.add(item.getId()));
+            return success(roleService.remove(idList));
+        } else {
+            throw new IllegalArgumentException("未传递要删除的角色记录");
+        }
     }
 
 }
